pyproject.toml
[project]
name = "zkb"
version = "0.1.0"
description = "A Zettelkasten-inspired knowledge base"
authors = [{ name = "Donald Thompson", email = "witt3rd@witt3rd.com" }]
dependencies = [
    "loguru>=0.7.2",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "qa-store>=0.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.2",
    "ruff>=0.5.0",
    "pre-commit>=3.7.1",
    "pip>=24.1.1",
    "ipykernel>=6.29.5",
]

[tool.hatch.build]
exclude = [".rye"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/zkb"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[project.urls]
GitHub = "https://github.com/witt3rd/zkb"
